增加定时任务业务实现

+++ property-dao/src/main/java/cn/zut/dao/entity/BusinessHouseRentEntity.java
@@ -33,7 +33,7 @@ public class BusinessHouseRentEntity {
     /**
      * 房屋面积
      */
-    private Double area;
+    private BigDecimal area;
     /**
      * 房屋地址
      */
